move DVC docker tests to new tests module
